Move the Protobuf-Java FQN cache to a separate class.
