+++ protobuf-plugin/plugin/src/main/java/org/spine3/gradle/protobuf/failures/MessageTypeCache.java
@@ -1,0 +1,108 @@
+/*
+ *
+ * Copyright 2016, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ */
+package org.spine3.gradle.protobuf.failures;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.protobuf.DescriptorProtos.DescriptorProto;
+import com.google.protobuf.DescriptorProtos.EnumDescriptorProto;
+import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
+import com.google.protobuf.DescriptorProtos.FileOptions;
+import org.spine3.gradle.protobuf.util.JavaCode;
+
+import java.util.Map;
+
+import static com.google.common.collect.Maps.newHashMap;
+
+/**
+ * A cache for the Protobuf message types parsed into appropriate Java types during the Spine Failure generation.
+ *
+ * @author Alexander Yevsyukov
+ * @author Alex Tymchenko
+ * @see FailuresGenPlugin
+ */
+public class MessageTypeCache {
+
+    /** A map from Protobuf type name to Java class FQN. */
+    private final Map<String, String> cachedMessageTypes = newHashMap();
+
+    //It's fine, as we are caching multiple message types per file.
+    @SuppressWarnings("MethodWithMultipleLoops")
+    /* package */ void cacheTypes(FileDescriptorProto fileDescriptor) {
+        final FileOptions options = fileDescriptor.getOptions();
+
+        final String sourceProtoPackage = fileDescriptor.getPackage();
+        final String protoPackage = !sourceProtoPackage.isEmpty()
+                                    ? (sourceProtoPackage + '.')
+                                    : "";
+        final String sourceJavaPackage = options.getJavaPackage();
+        String javaPackage = !sourceJavaPackage.isEmpty()
+                                   ? sourceJavaPackage + '.'
+                                   : "";
+
+        if (!options.getJavaMultipleFiles()) {
+            final String singleFileSuffix = JavaCode.getOuterClassName(fileDescriptor);
+            javaPackage = javaPackage + singleFileSuffix + '.';
+        }
+
+        for (DescriptorProto msg : fileDescriptor.getMessageTypeList()) {
+            cacheMessageType(msg, protoPackage, javaPackage);
+        }
+
+        for (EnumDescriptorProto enumType : fileDescriptor.getEnumTypeList()) {
+            cacheEnumType(enumType, protoPackage, javaPackage);
+        }
+    }
+
+    /**
+     * Obtain an immutable copy of the parsed Protobuf type names mapped to FQN of related Java classes.
+     *
+     * @return current cache contents
+     */
+    /* package */ Map<String, String> getCachedTypes() {
+        final ImmutableMap<String, String> immutable = ImmutableMap.copyOf(cachedMessageTypes);
+        return immutable;
+    }
+
+    //It's fine, as we are caching multiple nested types per message descriptor.
+    @SuppressWarnings("MethodWithMultipleLoops")
+    private void cacheMessageType(DescriptorProto msg, String protoPrefix, String javaPrefix) {
+        final String key = protoPrefix + msg.getName();
+        final String value = javaPrefix + msg.getName();
+        cachedMessageTypes.put(key, value);
+        if (msg.getNestedTypeCount() > 0 || msg.getEnumTypeCount() > 0) {
+            final String nestedProtoPrefix = protoPrefix + msg.getName() + '.';
+            final String nestedJavaPrefix = javaPrefix + msg.getName() + '.';
+            for (DescriptorProto nestedMsg : msg.getNestedTypeList()) {
+                cacheMessageType(nestedMsg, nestedProtoPrefix, nestedJavaPrefix);
+            }
+            for (EnumDescriptorProto enumType : msg.getEnumTypeList()) {
+                cacheEnumType(enumType, nestedProtoPrefix, nestedJavaPrefix);
+            }
+        }
+    }
+
+    private void cacheEnumType(EnumDescriptorProto descriptor, String protoPrefix, String javaPrefix) {
+        final String key = protoPrefix + descriptor.getName();
+        final String value = javaPrefix + descriptor.getName();
+        cachedMessageTypes.put(key, value);
+    }
+}
